Refactored naming and initialized files for pose detection app
