--- src/components/portfolio/webcam-detection/index.tsx
@@ -1,7 +1,0 @@
-import Camera from "./Camera";
-
-const WebcamDetectionComponent = () => {
-	return <Camera />;
-};
-
-export default WebcamDetectionComponent;
