+++ src/components/portfolio/webcam-object-detection/index.tsx
@@ -1,0 +1,7 @@
+import Camera from "./Camera";
+
+const WebcamObjectDetectionComponent = () => {
+	return <Camera />;
+};
+
+export default WebcamObjectDetectionComponent;
