+++ src/features/webcam-object-detection/const.ts
