+++ src/features/webcam-object-detection/model.ts
