+++ src/features/webcam-object-detection/types.ts
