+++ src/pages/webcam-object-detection.astro
@@ -1,18 +1,18 @@
 ---
 import PageLayout from "@/layouts/PageLayout.astro";
 import ContainerSection from "@/components/common/ContainerSection.astro";
-import WebcamDetectionComponent from "@/components/portfolio/webcam-detection";
+import WebcamObjectDetectionComponent from "@/components/portfolio/webcam-object-detection";
 import HeaderPortfolio from "@/components/portfolio/common/header-portfolio";
 ---
 
-<PageLayout meta={{ title: "Webcam Detection App" }}>
+<PageLayout meta={{ title: "Webcam Object Detection App" }}>
 	<ContainerSection maxWidth="max-w-2xl">
 		<HeaderPortfolio
 			client:visible
-			title={"Webcam Detection"}
+			title={"Webcam Object Detection"}
 			description={"A real-time image object detection application using your webcam. This project leverages the power of TensorFlow.js to detect and identify objects in your webcam feed."}
-			chipsList={["React", "Typescript", "react-webcam", "Coco-ssd", "Tensorflow.js"]}
+			chipsList={["Typescript", "react-webcam", "Coco-ssd model", "Tensorflow.js"]}
 		/>
-		<WebcamDetectionComponent client:visible />
+		<WebcamObjectDetectionComponent client:visible />
 	</ContainerSection>
 </PageLayout>
