+++ src/pages/webcam-pose-detection.astro
@@ -1,0 +1,18 @@
+---
+import PageLayout from "@/layouts/PageLayout.astro";
+import ContainerSection from "@/components/common/ContainerSection.astro";
+import WebcamPoseDetectionComponent from "@/components/portfolio/webcam-object-detection";
+import HeaderPortfolio from "@/components/portfolio/common/header-portfolio";
+---
+
+<PageLayout meta={{ title: "Webcam Pose Detection App" }}>
+	<ContainerSection maxWidth="max-w-2xl">
+		<HeaderPortfolio
+			client:visible
+			title={"Webcam Detection"}
+			description={"A real-time image object detection application using your webcam. This project leverages the power of TensorFlow.js to detect and identify objects in your webcam feed."}
+			chipsList={["React", "Typescript", "react-webcam", "Coco-ssd", "Tensorflow.js"]}
+		/>
+		<WebcamPoseDetectionComponent client:visible />
+	</ContainerSection>
+</PageLayout>
